+++ MongoMS.ServerExplorer.Addin/ViewModel/CollectionViewModel.cs
@@ -1,0 +1,32 @@
+using System.Collections.ObjectModel;
+using Microsoft.Practices.Unity;
+using MongoDB.Driver;
+using MongoMS.Common;
+
+namespace MongoMS.ServerExplorer.Addin.ViewModel
+{
+    public class CollectionViewModel
+    {
+        private readonly MongoCollection _collection;
+        private readonly IUnityContainer _unity;
+
+        public CollectionViewModel(MongoCollection collection, IUnityContainer unity)
+        {
+            _collection = collection;
+            _unity = unity;
+        }
+
+        public object CmdParameter
+        {
+            get { return _collection; }
+        }
+        public string Name { get { return _collection.Name; } }
+        public ObservableCollection<IMenuCommand> Menu
+        {
+            get
+            {
+                return _unity.Resolve<ObservableCollection<IMenuCommand>>(ContextMenuLevel.Collection.ToString());
+            }
+        }
+    }
+}
